pingerd for dhmon
Pingerd is a small C daemon that listens on two ZMQ channels.

The first channel :5560 is a subscriber. Push a string with an IP in it
and it will send an ICMP request to it.

The second channe :5561 is a publisher. Every ICMP reply will be
published on this channel.

pinger.py takes one argument: the IP to ping and will then ping it as
fast as it can. It's just an example of how to talk to pingerd.