Speedup tweaks and switch to PyPy
In the last commit we hit some kind of locality issue I believe
because the whole exporter craweled to a 10x slowdown.

Change to compute the output when the copy is made and use PyPy
for the exporter. PyPy rocks, the speedup is amazing.

Since I don't have netsnmp for pypy and a simple pip install doesn't
cut it for netsnmp, we have to delay the loading of netsnmp.