Introduce backend cached metrics
As all metrics in analytics are stateless we introduce a cache that will
produce up to CACHE_TIME stale metrics. This has the benefit of making all
client requests very fast, at the cost of being slightly stale.

This reduces the load time of dhmap from 5s to about 1s.